Day 5: Mesh shaders!
Today we've written our first mesh shader that uses meshlet data
prepared on the CPU. We are *not* utilizing the GPU efficiently yet as
our mesh shader is only using one thread on the warp, which should run
at ~3% utilization, and we are *not* optimizing the mesh well which
reflects poorly on the number of meshlets. We will fix this tomorrow.

We also hit two other bugs (possibly) - one in glslang that may not be a
bug but it's a weird behavior where it adds a shader capability we
aren't even using, and one definite bug in validation layers where they
don't correctly check that the buffer can be used as a storage buffer.