Day 8: Cone culling
We have implemented meshlet cone computation and we can successfully
cull entire meshlets in the mesh shader based on this cone.

Culling efficiency is reasonable on some meshes but deteriorates on
really low poly meshes or really high poly meshes. Which seems expected
as it's just a limitation of the cluster culling technique.