Repack culling data
We don't need full frustum equations for symmetric frustums - in fact,
we can utilize the symmetry to substantially simplify the plane checks.
Only one comparison is required for a pair of symmetric planes, and two
out of four coefficients are zero so we don't need to store them or use
them for comparison.

Additionally refactor LOD computation to be easily controllable from the
CPU side using a basic base*pow(step,i) exponential formula.