shaders: Fix culling for rotated meshes
Whenever mesh bounding center is not 0,0,0, we need to rotate it around
the mesh origin to actually transform the bounding sphere to world/view
space. Failure to do so results in incorrect culling.

Also extract drawId into a variable in all shaders to make it easier to
compare against/debug.

Fixes #5.