Do not request debug layers when they are not present
Instead of unconditionally enabling debug layers in Debug, we check if
they are supported first. This makes it easier to run niagara on systems
without Vulkan SDK installed, but with a functioning Vulkan runtime and
headers.

We also decouple VK_EXT_debug_report which, for now, is still enabled
unconditionally as drivers seem to support it. This can be changed
separately.