Fix occlusion culling to be conservative
floor(log2) is incorrect; the correct version is ceil(log2). When using
floor, we essentially ensure that the rectangle is between 1x1 and 2x2
pixels in the mip space. ceil ensures that the rectangle is 1x1 or
smaller.

The issue is that because the rectangle may be offset arbitrarily on the
grid, a 2x2 rectangle covers a 3x3 pixel area and as such requires more
than one bilinear texture fetch to correctly reject. ceil solves this
problem at the expense of some loss of occlusion efficiency.

If desired, this can be balanced by using floor() and sampling 4 corners
of AABB which should be conservative at a small extra cost.