Adjust settings for better performance on RDNA3
Triangle culling seems fairly unnecessary on both NV and RDNA3; this
change also rebalances meshlets to use a configuration that is more
optimal for RDNA3 and NV, but less optimal for RDNA2 due to export
restrictions.

Also, we are running into some sort of performance bottleneck on RDNA3
with cluster occlusion culling that makes the second pass *very*
expensive despite the fact that it's mostly just confirming cluster
visibility; for now disable it by default to avoid surprises.