Use normals for LODs and steeper triangle count curve
Ideally we would actually dynamically vary the error, not the triangle
count, when generating LODs, but the rendering code will now just work
so keep it as is for now with a steeper curve: 0.75^n is not steep
enough to select the optimal LODs at large distances when we only have 8
LODs at most.

We now also include the normals into the LOD calculation which fixes the
normal popping when LOD switches close by. This technically requires
meshoptimizer 0.22 to work perfectly but that version isn't tagged so
we'll bump it separately.