Day 21: Heroically implement cluster culling ourselves
Because of AMD CP bottleneck that we are hitting again, but this time in
the task shader emulation (async compute work that driver is submitting
is stalling the geometry work in render early, and is heavily
bottlenecked by the serial CP processing in render late); this is
significantly affecting the performance of meshlet culling pipeline in
all configurations, even ones where cluster occlusion is disabled.

Fixing this required us to run cluster culling in compute vs task
ourselves, fill a cluster index buffer (4b per visible cluster overhead)
and use mesh shaders without task shaders.

On 7900 GRE, this is dramatically speeding up our pipeline: 10ms to
1.4ms with cluster culling enabled, 5ms to 3.4ms with cluster culling
disabled.