Adjust the cluster limits to be 4x higher
We now have enough buffer space for 16M meshlets ~= 1.5B triangles to
render in a single frame; this requires a slightly larger grid size,
which doesn't seem to penalize performance.

Also any excess here or in task submit is correctly clamped to the
buffer limit to avoid overruns on padding.