Refactor task/mesh shader split to use the same mesh shader source
We *can* use specialization constants it looks like! The spec doesn't
mandate that the payload is used iff task shader is linked in, so
presumably as long as we don't read the payload we're okay.

This alone is sufficient to remove the extra shader source; this change
also cleans up C++ source with saner variable names and restructures the
payload so that it uses the same 32-bit command-cluster indices to make
the code almost the same between the two.