Fix 2D memory layout to be more efficient to access
We used XxK instead of KxX when unwrapping a large 1D array; if we
assume that the workgroup dispatch is roughly sequential, this results
in suboptimal memory access patterns.

This is significantly better on RDNA iGPU (and fixes the regression wrt
task shading path); it is a little worse on 7900 GRE, but that's a
problem for another day.