Clean up legacy extension fallbacks
We added KHR_push_descriptor fallback to work around AMD official
drivers being unwilling to support this. This has changed since, and it
would be cleaner to rely on this feature unconditionally; we had some
comments demonstrating direct extension usage but they were wrong.

Also remove profiling support; this was only added to fix clocks for
RDNA2, but radv no longer exposes this on RDNA3 and the extension in
general isn't very popular.