Basic support for macOS compilation
MoltenVK seems to be getting a little closer to the minimal feature set.
While it's still not quite there, there's a couple simple fixes we can do
to make it possible to compile.

Note: since we need 1.3, this is not enough to create the device, so the
swapchain code is unfinished.

Code changes partially taken from https://github.com/zeux/niagara/pull/37