Add VK_CHECK_FORCE for critical errors
In Release, we elide assertions for all other APIs, but a GPU hang might
be more recoverable if we actually abort earlier instead of submitting
more GPU work that creates a cascade of issues. Might be better to
always check everything in Release but for now add a separate call.