Switch to float16 positions
In theory float16 may not be enough for some large meshes, but for now
just switch to it unconditionally; do this in a separate commit so that
we can revert it later if need be.

This reduces the vertex size from 24 bytes to 20 bytes; we can
separately go down to 16 bytes by packing TBN better.