Stop overallocating buffers for scene data
In the future we will probably merge a lot of these into larger buffers,
but for now we can correctly size the buffers we allocate; on Bistro
this reduces our geometry requirements from 1.4 GB to 0.7 GB (with ~half
of the memory required by BLAS...).

We now also print the total texture memory requirements, which are ~2 GB
for Bistro.