Implement shadow checkerboard
This is a variant of the experiment on-stream, but somewhat less
aggressive: we trace half of the rays in a checkerboard pattern, and
reconstruct missing pixels from 4 neighbors using bilateral weights.

This also results in increased aliasing but it's less severe vs
half-res, and of course the performance gains are less substantial.