Day 28: Implement initial support for glTF animations!
We currently only handle animations of leaf nodes and expect uniform
T/R/S tracks with the same frequency, and using linear interpolation
only. This is to avoid the need to build a full persistent scene graph
with hierarchical transforms et al.

We update the draw data for every object by writing directly to GPU
memory; in theory some drivers may not expose this, but if this is
actually a blocker for us we will fix this.

Also for now work around a query not_ready status on AMDVLK; need to
debug that separately in the future.