Use 16-bit vertex references for meshlets
Most meshlets reference a short range of vertices which can fit into 16
bits with an offset. Since we already store base vertex per meshlet this
can be very easily supported with a conditional load in the shader.

This reduces meshlet data from 35 MB to 25 MB on Bistro.

After this, 20% of meshlets can still use 1-byte range but the savings
are less pronounced so this can be done at some future point separately.