Day 29: Implement support for multiple frames in flight
For this to work we duplicate command pools/buffers, query pools,
semaphores and fences. There might be cleaner solutions for pools in the
future.

Note that there's a race condition with respect to transform update that
we need for animations; we are going to fix that separately / in the
future.