Fix release semaphore usage
Latest Vulkan SDK added a new validation error that claims that using
release semaphores from 2 frames ago with a 3-image swapchain is unsafe,
as the release semaphore may have not been signaled if the presentation
queue is separate.

Fix this by tying the release (present) semaphores to the image index
instead of frame index. Not sure if this is truly necessary but this
fixes the validation error at least.