Day 30: Start implementing stage barriers
Instead of resource barriers that synchronize (or transition) individual
buffers and images, we are starting to migrate to memory barriers,
abstracted behind a very nice and clean stageBarrier function. The hope
is that this allows us to express the same constraints to the driver at
the same resulting performance level, but makes things much easier to
manage if we change the rendering code.