Fix meshlets overallocation during scene loading
This was accidentally committed at some point during CLRT adjustments;
as a result, on large models we have way too many meshlets allocated,
which results in way too many vertices/triangles; this needlessly slows
generation down.

These bounds can be even tighter in the future, but for now it's
simplest to leave as is.