Switch to dual X11/Wayland build with runtime detection
Instead of using compile-time detection of Wayland vs X11, we initialize
either extension if supported, and use the platform GLFW selects
automatically to initialize the window. This will default to Wayland if
supported.

Additionally, we now support pulling glfw3 from CMake system modules,
which removes the need to compile it on Linux (as that requires
installing a bunch of odd dependencies).